refactor: moving test from `shared` -> `spyglass-lens`
