+++ crates/shared/src/config.rs
@@ -291,25 +291,3 @@ impl Config {
         config
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::LensConfig;
-
-    #[test]
-    fn test_into_regexes() {
-        let config = LensConfig {
-            domains: vec!["paulgraham.com".to_string()],
-            urls: vec!["https://oldschool.runescape.wiki/wiki/".to_string()],
-            ..Default::default()
-        };
-
-        let regexes = config.into_regexes();
-        dbg!(&regexes);
-        assert_eq!(regexes.len(), 2);
-        // Should contain domain regex
-        assert!(regexes.contains(&"^(http://|https://)paulgraham\\.com.*".to_string()));
-        // Should contain url prefix regex
-        assert!(regexes.contains(&"^https://oldschool.runescape.wiki/wiki/.*".to_string()));
-    }
-}
